refactor: move FileEntry to model package
