+++ iomete-lakehouse-backup/src/main/kotlin/com/iomete/backup/model/Types.kt
@@ -1,8 +1,13 @@
-@file:Suppress("ktlint:standard:filename")
-
 package com.iomete.backup.model
 
-import com.iomete.backup.fs.FileEntry
+import java.io.Serializable
+
+// Serializable so instances distribute across Spark executors via RDD.
+data class FileEntry(
+    val path: String,
+    val size: Long,
+    val modificationTime: Long,
+) : Serializable
 
 data class SourceListing(
     val files: List<FileEntry>,
